// lib/api/generated/models/models-case-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ModelsCaseUserSummary } from './models-case-user-summary';

export interface ModelsCaseResponse {
    'assignee'?: ModelsCaseUserSummary;
    'assignee_id'?: string;
    'case_number'?: string;
    'created_at'?: string;
    'description'?: string;
    'id'?: string;
    'priority'?: string;
    'related_entity_id'?: string;
    'related_entity_type'?: string;
    'reporter'?: ModelsCaseUserSummary;
    'reporter_id'?: string;
    'resolved_at'?: string;
    'status'?: string;
    'title'?: string;
    'updated_at'?: string;
}